requirements.txt
pyaes==1.6.1
pyasn1==0.6.1
python-dotenv==1.0.1
rsa==4.9
Telethon==1.38.1
